+++ src/main/java/com/coniverse/dangjang/domain/guide/bloodsugar/guideformat/GuideFormat.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.guide.bloodsugar.enums;
+package com.coniverse.dangjang.domain.guide.bloodsugar.guideformat;
 
 public interface GuideFormat {
 	String getTitle();
