+++ src/main/java/com/coniverse/dangjang/domain/guide/bloodsugar/service/BloodSugarGuideGenerateService.java
@@ -11,12 +11,12 @@ import com.coniverse.dangjang.domain.code.enums.CommonCode;
 import com.coniverse.dangjang.domain.code.enums.GroupCode;
 import com.coniverse.dangjang.domain.guide.bloodsugar.document.BloodSugarGuide;
 import com.coniverse.dangjang.domain.guide.bloodsugar.document.SubGuide;
-import com.coniverse.dangjang.domain.guide.bloodsugar.enums.CautionGuideFormat;
-import com.coniverse.dangjang.domain.guide.bloodsugar.enums.GuideFormat;
-import com.coniverse.dangjang.domain.guide.bloodsugar.enums.HypoglycemiaGuideFormat;
-import com.coniverse.dangjang.domain.guide.bloodsugar.enums.HypoglycemiaSuspectGuideFormat;
-import com.coniverse.dangjang.domain.guide.bloodsugar.enums.NormalGuideFormat;
-import com.coniverse.dangjang.domain.guide.bloodsugar.enums.WarningGuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.CautionGuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.GuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.HypoglycemiaGuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.HypoglycemiaSuspectGuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.NormalGuideFormat;
+import com.coniverse.dangjang.domain.guide.bloodsugar.guideformat.WarningGuideFormat;
 import com.coniverse.dangjang.domain.guide.bloodsugar.mapper.BloodSugarGuideMapper;
 import com.coniverse.dangjang.domain.guide.bloodsugar.repository.BloodSugarGuideRepository;
 import com.coniverse.dangjang.domain.guide.common.dto.GuideResponse;
